tests/wpt/tests/css/css-contain/content-visibility/content-visibility-hidden-offsetTop-left-width-height.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Content Visibility: offsetLeft/offsetTop/offsetWidth/offsetHeight measure correctly</title>
<link rel="help" href="https://drafts.csswg.org/css-contain/#content-visibility">
<meta name="assert" content="offsetLeft/offsetTop/offsetWidth/offsetHeight measure correctly in content-visibility hidden subtree">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
body {
  margin: 0;
  padding: 0;
}
#outer {
  width: 100px;
  height: 100px;
  background: lightblue;

  content-visibility: hidden;
}
#inner {
  border: 25px black solid;
  margin: 25px;
  width: 50px;
  height: 50px;
  background: lightgreen;
}
</style>

<body>
<div id="outer"><div id="inner"></div></div>
</body>

<script>
test(() => {
  const inner = document.getElementById("inner");

  inner.style.marginLeft = "20px";
  assert_equals(inner.offsetLeft, 20, "left");

  inner.style.marginTop = "20px";
  assert_equals(inner.offsetTop, 20, "top");

  inner.style.width = "40px";
  assert_equals(inner.offsetWidth, 90, "width");

  inner.style.height = "40px";
  assert_equals(inner.offsetHeight, 90, "height");
});

</script>
</html>
